modify rotation logic, index of body item does not impact rotation
