+++ client/components/Main/Project/index.jsx
@@ -1,131 +1,37 @@
 import React, { Component } from 'react'
 import { connect } from 'react-redux'
-import { TransitionGroup } from 'react-transition-group'
-import Draggable from 'gsap/Draggable'
-import { TweenMax } from 'gsap'
-import { Slide } from '../../Shared/Transition'
-import { MainContainer } from '../../Shared/Styles'
-import Navigation from '../Navigation'
 import ProjectView from './ProjectView'
+import { TransitionGroup } from 'react-transition-group'
+import { Slide } from '../../Shared/Transition'
+import BodyComponent from '../Body'
 
-import { rotationChange } from '../../../reducers/events'
+import { viewData } from '../../Shared/Data'
 
-const Project = ({  direction, inputBody, inputMain, inputNav, language, navigationList, targetOffset, viewIndex }) => {
-  const currentView = navigationList[viewIndex]
-
+const Project = ({ navigationList }) => {
   return (
-    <MainContainer innerRef={inputMain}>
-      <div style={{maxHeight: '100px'}}>
-        <Navigation
-          navigationList={navigationList}
-          currentIndex={viewIndex}
-          getDom={inputNav}
-        />
-      </div>
-      <TransitionGroup>
-        <Slide key={viewIndex} targetOffset={targetOffset} exit={false}>
-          <div ref={inputBody}>
-            <ProjectView viewIndex={viewIndex} language={language} />
-          </div>
-        </Slide>
-      </TransitionGroup>
-    </MainContainer>
+    <BodyComponent
+      navigationList={navigationList}
+    >
+      <ProjectView
+      />
+    </BodyComponent>
   )
 }
 
 class LocalContainer extends Component {
-  constructor(props) {
-    super(props)
-    this.state = {
-      navigationList: ['AUDIOSPHERE', 'STACKQUEST', 'PORTFOLIO'],
-      direction: 'right',
-      targetOffset: 20,
-    }
+  static get navigationList () {
+    return viewData.projects.navigationList
   }
 
-  static getDirection(magnitude) {
-    return magnitude > 0 ? 'left' : 'right'
-  }
-
-  static nearest(fn, ratio) {
-    return function (getRatio) {
-      return function () {
-        const { rotation } = this
-            , targetRotation = getRatio ? Math.round(rotation / ratio) * ratio : rotation
-        fn(targetRotation)
-      }
-    }
-  }
-
-  static slide(targetDOM, length) {
-    const ratio = (targetDOM.offsetWidth / 2) / (360 / length) // get the half of targetdom
-        , getOffset = ((rat) => (rotation) => (rotation * rat))(ratio)
-    let prevRotation = 0
-
-    return (magnitude) => {
-      const targetOffset = getOffset(prevRotation -= magnitude)
-
-      TweenMax.to(targetDOM, 0.7, {
-        marginRight: `${targetOffset}px`
-      }) // separate this
-
-      return targetOffset
-    }
-  } // for gradual slide
-
-  componentDidMount() {
-    const { length } = this.state.navigationList
-
-    this.slideBody = LocalContainer.slide(this.body, length)
-    this.getTargetRotation = LocalContainer.nearest(this.props.rotationChange, 360 / length)
-
-    // DEFINE DRAGGABLE
-    Draggable.create(this.nav, {
-      type: 'rotation',
-      trigger: this.mainDiv,
-      onDrag: this.getTargetRotation(false),
-      onDragEnd: this.getTargetRotation(true),
-    })
-  }
-
-  componentWillReceiveProps({ rotation }) {
-    const magnitude = this.props.rotation - rotation
-        , targetOffset = this.slideBody(magnitude) // get magnitude
-        , direction = LocalContainer.getDirection(magnitude)
-    this.setState(Object.assign({}, this.state, {direction, targetOffset}))
-  }
-
-  shouldComponentUpdate({ viewIndex }) {
-    return this.props.viewIndex !== viewIndex
-  }
-
-  componentDidUpdate() {
-    // reassign slideBody with new props index
-    this.slideBody = LocalContainer.slide(this.body, this.state.navigationList.length, this.props.viewIndex)
-  }
-
-  render() {
+  render () {
 
     return (
       <Project
-        {...this.props}
-        {...this.state}
-        inputBody={div => this.body = div}
-        inputMain={div => this.mainDiv = div}
-        inputNav={div => this.nav = div}
+        navigationList={LocalContainer.navigationList}
       />
     )
   }
 }
 
-const mapStateToProps = (state) => ({
-  language: state.language.language,
-  viewIndex: state.events.viewIndex,
-  rotation: state.events.rotation
-})
 
-const mapDispatchToProps = (dispatch) => ({
-  rotationChange: (rotation) => dispatch(rotationChange(rotation)),
-})
-
-export default connect(mapStateToProps, mapDispatchToProps)(LocalContainer)
+export default LocalContainer
