+++ client/reducers/events.js
@@ -26,7 +26,7 @@ export default (state = initialState, action) => {
     return Object.assign({}, state, {loaded: true})
 
   case ROTATION_CHANGE:
-    return Object.assign({}, state, {rotation: action.rotation % 360}) // patch it up
+    return Object.assign({}, state, {rotation: action.rotation}) // patch it up
 
   case ROTATION_RESTART:
     return Object.assign({}, state, {rotation: 0})
